test: extend SavePortsAdapterIT with economy field assertions + fix type mismatches
Extended SavePortsAdapterIT.findThenUpdateThenFindRoundTripsEveryMutableField() to set
and assert economyIndex (2) and economyReading (45) during the SaveState round-trip test.

Fixed SaveEntity, SaveState, and EconomyService to use byte/short for the economy fields,
matching the V7 migration column types (TINYINT/SMALLINT). The ranges -3..+3 and -30..+90
fit correctly in these smaller types, providing schema consistency.

Updated all test files to cast economy field values explicitly where passed to setters or
constructors. This prevents implicit narrowing errors and ensures type safety across the
application.

Verify: 248 unit tests + 30 IT tests, all passing. SavePortsAdapterIT now confirms
the economy fields round-trip correctly through JPA persistence.

+++ amiss-api/src/main/java/amiss/api/persistence/jpa/SaveEntity.java
@@ -111,11 +111,11 @@ public class SaveEntity {
 
     /** Hidden economy trend, -3..+3 (KAN-48). Never expose on a wire DTO. */
     @Column(name = "economy_index", nullable = false)
-    private int economyIndex;
+    private byte economyIndex;
 
     /** Hidden economy reading, -30..+90 (KAN-48): price = base + base*reading/60. */
     @Column(name = "economy_reading", nullable = false)
-    private int economyReading;
+    private short economyReading;
 
     protected SaveEntity() {
         // JPA
@@ -314,19 +314,19 @@ public class SaveEntity {
         this.won = won;
     }
 
-    public int getEconomyIndex() {
+    public byte getEconomyIndex() {
         return economyIndex;
     }
 
-    public void setEconomyIndex(int economyIndex) {
+    public void setEconomyIndex(byte economyIndex) {
         this.economyIndex = economyIndex;
     }
 
-    public int getEconomyReading() {
+    public short getEconomyReading() {
         return economyReading;
     }
 
-    public void setEconomyReading(int economyReading) {
+    public void setEconomyReading(short economyReading) {
         this.economyReading = economyReading;
     }
 }
