+++ amiss-api/src/test/java/amiss/api/web/SecurityRulesTest.java
@@ -109,7 +109,7 @@ class SecurityRulesTest {
     @Test
     void authenticatedAsAlice_readingOwnSave_is200() throws Exception {
         SaveState save = new SaveState(7L, "alice", "My Save", 0, 2, 3600, 1, 120, 0, 0, 1, 1, 1,
-                null, 0, 10, 20, null, 0, 200, 100, 30, 50, false, 0, 0);
+                null, 0, 10, 20, null, 0, 200, 100, 30, 50, false, (byte) 0, (short) 0);
         when(scope.require(eq(7L), any())).thenReturn(save);
         when(assembler.assemble(services, save)).thenReturn(dto());
